*** empty log message ***
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main/sip@272 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resiprocate/os/Log.cxx
@@ -150,7 +150,7 @@ Log::timestamp()
    char msbuf[5];
    /* Dividing (without remainder) by 1000 rounds the microseconds
       measure to the nearest millisecond. */
-   sprintf(msbuf, ".%3.3ld", (tv.tv_usec / 1000));
+   sprintf(msbuf, ".%3.3ld", long(tv.tv_usec / 1000));
    
    int datebufCharsRemaining = DATEBUF_SIZE - strlen (datebuf);
    strncat (datebuf, msbuf, datebufCharsRemaining - 1);
